enum クラスへの書き換え
