--- app/src/main/java/jp/co/yumemi/droidtraining/presentation/constant/Image/imageConstants.kt
@@ -1,16 +1,0 @@
-package jp.co.yumemi.droidtraining.presentation.constant.Image
-
-import android.annotation.SuppressLint
-import androidx.compose.ui.graphics.Color
-import jp.co.yumemi.droidtraining.R
-
-
-class ImageConstants {
-    companion object {
-        val sunny: Int = R.drawable.sunny
-        val cloudy: Int = R.drawable.cloudy
-        val rainy: Int = R.drawable.rainy
-        val snow: Int = R.drawable.snow
-    }
-}
-
